update docs, clean up
