+++ e2e/utils/_.bash
@@ -323,7 +323,6 @@ use_default_wallet_wasm() {
 
 get_webserver_port() {
   dfx info webserver-port
-  # cat "$E2E_NETWORK_DATA_DIRECTORY/webserver-port"
 }
 overwrite_webserver_port() {
   echo "$1" >"$E2E_NETWORK_DATA_DIRECTORY/webserver-port"
