+++ src/dfx/src/commands/info/webserver_port.rs
@@ -1,0 +1,16 @@
+use crate::lib::provider::{create_network_descriptor, LocalBindDetermination};
+use crate::{DfxResult, Environment};
+
+pub(crate) fn get_webserver_port(env: &dyn Environment) -> DfxResult<String> {
+    let port = create_network_descriptor(
+        env.get_config(),
+        env.get_networks_config(),
+        None,
+        None,
+        LocalBindDetermination::ApplyRunningWebserverPort,
+    )?
+    .local_server_descriptor()?
+    .bind_address
+    .port();
+    Ok(format!("{}", port))
+}
